Page management to main frontend (#258)
* Remove old and unused page creation code

* Move page management to main frontend
