+++ services/cms/src/services/backend/pages/index.ts
@@ -1,16 +1,9 @@
-import { NewPage, Page, PageUpdate } from "../../../shared-module/bindings"
+import { Page, PageUpdate } from "../../../shared-module/bindings"
 import { cmsClient } from "../cmsClient"
 
 export const fetchPageWithId = async (pageId: string): Promise<Page> => {
   const data = (await cmsClient.get(`/pages/${pageId}`, { responseType: "json" })).data
   return data
-}
-
-export const postNewPage = async (data: NewPage): Promise<Page> => {
-  const response = await cmsClient.post("/pages", data, {
-    headers: { "Content-Type": "application/json" },
-  })
-  return response.data
 }
 
 export const updateExistingPage = async (
@@ -24,10 +17,5 @@ export const updateExistingPage = async (
       headers: { "Content-Type": "application/json" },
     },
   )
-  return response.data
-}
-
-export const deletePage = async (page_id: string): Promise<Page> => {
-  const response = await cmsClient.delete(`/pages/${page_id}`)
   return response.data
 }
