+++ services/headless-lms/src/controllers/main_frontend/courses.rs
@@ -1,16 +1,19 @@
 //! Controllers for requests starting with `/api/v0/main-frontend/courses`.
 use crate::{
-    controllers::ControllerResult,
+    controllers::{helpers::media::upload_media_for_course, ControllerResult, UploadResult},
     domain::authorization::AuthUser,
     models::{
         course_instances::CourseInstance,
-        courses::{Course, CourseUpdate, NewCourse},
+        courses::{Course, CourseStructure, CourseUpdate, NewCourse},
         exercises::Exercise,
         submissions::{SubmissionCount, SubmissionCountByExercise, SubmissionCountByWeekAndHour},
     },
+    utils::file_store::FileStore,
+    ApplicationConfiguration,
 };
-use actix_web::web::ServiceConfig;
+use actix_multipart as mp;
 use actix_web::web::{self, Json};
+use actix_web::{web::ServiceConfig, HttpRequest};
 use sqlx::PgPool;
 use uuid::Uuid;
 
@@ -149,6 +152,110 @@ async fn delete_course(
     let mut conn = pool.acquire().await?;
     let course = crate::models::courses::delete_course(&mut conn, *request_course_id).await?;
     Ok(Json(course))
+}
+
+/**
+GET `/api/v0/main-frontend/courses/:course_id/structure` - Returns the structure of a course.
+# Example
+```json
+{
+  "course": {
+    "id": "d86cf910-4d26-40e9-8c9c-1cc35294fdbb",
+    "slug": "introduction-to-everything",
+    "created_at": "2021-04-28T10:40:54.503917",
+    "updated_at": "2021-04-28T10:40:54.503917",
+    "name": "Introduction to everything",
+    "organization_id": "1b89e57e-8b57-42f2-9fed-c7a6736e3eec",
+    "deleted_at": null
+  },
+  "pages": [
+    {
+      "id": "f3b0d699-c9be-4d56-bd0a-9d40e5547e4d",
+      "created_at": "2021-04-28T13:51:51.024118",
+      "updated_at": "2021-04-28T14:36:18.179490",
+      "course_id": "d86cf910-4d26-40e9-8c9c-1cc35294fdbb",
+      "content": [],
+      "url_path": "/",
+      "title": "Welcome to Introduction to Everything",
+      "deleted_at": null,
+      "chapter_id": "d332f3d9-39a5-4a18-80f4-251727693c37"
+    }
+  ],
+  "chapters": [
+    {
+      "id": "d332f3d9-39a5-4a18-80f4-251727693c37",
+      "created_at": "2021-04-28T16:11:47.477850",
+      "updated_at": "2021-04-28T16:11:47.477850",
+      "name": "The Basics",
+      "course_id": "d86cf910-4d26-40e9-8c9c-1cc35294fdbb",
+      "deleted_at": null,
+      "chapter_image_url": "http://project-331.local/api/v0/files/uploads/organizations/1b89e57e-8b57-42f2-9fed-c7a6736e3eec/courses/d86cf910-4d26-40e9-8c9c-1cc35294fdbb/images/mbPQh8th96TdUwX96Y0ch1fjbJLRFr.png",
+      "chapter_number": 1,
+      "front_page_id": null
+    }
+  ]
+}
+```
+*/
+#[instrument(skip(pool, file_store, app_conf))]
+async fn get_course_structure<T: FileStore>(
+    request_course_id: web::Path<Uuid>,
+    pool: web::Data<PgPool>,
+    user: AuthUser,
+    file_store: web::Data<T>,
+    app_conf: web::Data<ApplicationConfiguration>,
+) -> ControllerResult<Json<CourseStructure>> {
+    let mut conn = pool.acquire().await?;
+    let course_structure = crate::models::courses::get_course_structure(
+        &mut conn,
+        *request_course_id,
+        file_store.as_ref(),
+        app_conf.as_ref(),
+    )
+    .await?;
+    Ok(Json(course_structure))
+}
+
+/**
+POST `/api/v0/main-frontend/courses/:course_id/upload` - Uploads a media (image, audio, file) for the course from Gutenberg page edit.
+
+Put the the contents of the media in a form and add a content type header multipart/form-data.
+# Example
+
+Request:
+```http
+POST /api/v0/main-frontend/pages/d86cf910-4d26-40e9-8c9c-1cc35294fdbb/upload HTTP/1.1
+Content-Type: multipart/form-data
+
+BINARY_DATA
+```
+
+Response:
+```json
+{
+    "url": "http://project-331.local/api/v0/files/organizations/1b89e57e-8b57-42f2-9fed-c7a6736e3eec/courses/d86cf910-4d26-40e9-8c9c-1cc35294fdbb/images/iHZMHdvsazy43ZtP0Ea01sy8AOpUiZ.png"
+}
+
+```
+*/
+#[instrument(skip(payload, request, pool, file_store, app_conf))]
+async fn add_media_for_course<T: FileStore>(
+    request_course_id: web::Path<Uuid>,
+    payload: mp::Multipart,
+    request: HttpRequest,
+    pool: web::Data<PgPool>,
+    user: AuthUser,
+    file_store: web::Data<T>,
+    app_conf: web::Data<ApplicationConfiguration>,
+) -> ControllerResult<Json<UploadResult>> {
+    let mut conn = pool.acquire().await?;
+    let course = crate::models::courses::get_course(&mut conn, *request_course_id).await?;
+
+    let media_path =
+        upload_media_for_course(request.headers(), payload, &course, file_store.as_ref()).await?;
+    let download_url = file_store.get_download_url(&media_path.as_path(), app_conf.as_ref());
+
+    Ok(Json(UploadResult { url: download_url }))
 }
 
 /**
@@ -314,7 +421,7 @@ The name starts with an underline in order to appear before other functions in t
 
 We add the routes by calling the route method instead of using the route annotations because this method preserves the function signatures for documentation.
 */
-pub fn _add_courses_routes(cfg: &mut ServiceConfig) {
+pub fn _add_courses_routes<T: 'static + FileStore>(cfg: &mut ServiceConfig) {
     cfg.route("/{course_id}", web::get().to(get_course))
         .route("", web::post().to(post_new_course))
         .route("/{course_id}", web::put().to(update_course))
@@ -324,6 +431,14 @@ pub fn _add_courses_routes(cfg: &mut ServiceConfig) {
             web::get().to(get_daily_submission_counts),
         )
         .route("/{course_id}/exercises", web::get().to(get_all_exercises))
+        .route(
+            "/{course_id}/structure",
+            web::get().to(get_course_structure::<T>),
+        )
+        .route(
+            "/{course_id}/upload",
+            web::post().to(add_media_for_course::<T>),
+        )
         .route(
             "/{course_id}/weekday-hour-submission-counts",
             web::get().to(get_weekday_hour_submission_counts),
