+++ services/headless-lms/src/controllers/main_frontend/pages.rs
@@ -1,0 +1,123 @@
+//! Controllers for requests starting with `/api/v0/main-frontend/pages`.
+use crate::{
+    controllers::ControllerResult,
+    domain::authorization::AuthUser,
+    models::pages::{NewPage, Page},
+};
+use actix_web::web::ServiceConfig;
+use actix_web::web::{self, Json};
+use sqlx::PgPool;
+use uuid::Uuid;
+
+/**
+POST `/api/v0/main-frontend/pages` - Create a new page.
+
+Please note that this endpoint will change all the exercise and exercise task ids you've created. Make sure the use the updated ids from the response object.
+
+If optional property front_page_of_chapter_id is set, this page will become the front page of the specified course part.
+
+# Example:
+
+Request:
+```http
+POST /api/v0/main-frontend/pages HTTP/1.1
+Content-Type: application/json
+
+{
+  "content": [
+    {
+      "type": "x",
+      "id": "2a4e517d-a7d2-4d82-89fb-a1333d8d01d1"
+    }
+  ],
+  "url_path": "/part-2/best-page",
+  "title": "Hello world!",
+  "course_id": "10363c5b-82b4-4121-8ef1-bae8fb42a5ce",
+  "chapter_id": "2495ffa3-7ea9-4615-baa5-828023688c79"
+}
+```
+
+Response:
+```json
+{
+  "id": "d90bf7ab-181c-4aa2-a87e-5c28238cc67d",
+  "created_at": "2021-03-12T09:27:36.428501",
+  "updated_at": "2021-03-12T09:27:36.428501",
+  "course_id": "10363c5b-82b4-4121-8ef1-bae8fb42a5ce",
+  "content": [
+    {
+      "id": "18110110-d02a-4432-8cb9-084d0c63a524",
+      "type": "x"
+    }
+  ],
+  "url_path": "/part-2/best-page",
+  "title": "Hello world!",
+  "deleted_at": null,
+  "chapter_id": "2495ffa3-7ea9-4615-baa5-828023688c79",
+  "front_page_of_chapter_id": null
+}
+```
+
+*/
+#[instrument(skip(pool))]
+async fn post_new_page(
+    payload: web::Json<NewPage>,
+    pool: web::Data<PgPool>,
+    user: AuthUser,
+) -> ControllerResult<Json<Page>> {
+    let mut conn = pool.acquire().await?;
+    let new_page = payload.0;
+    let page = crate::models::pages::insert_page(&mut conn, new_page).await?;
+    Ok(Json(page))
+}
+
+/**
+DELETE `/api/v0/main-frontend/pages/:page_id` - Delete a page, related exercises, and related exercise tasks by id.
+
+
+# Example
+
+Request: `DELETE /api/v0/main-frontend/pages/40ca9bcf-8eaa-41ba-940e-0fd5dd0c3c02`
+
+Response:
+```json
+{
+  "id": "40ca9bcf-8eaa-41ba-940e-0fd5dd0c3c02",
+  "created_at": "2021-03-08T20:14:56.216394",
+  "updated_at": "2021-03-08T20:29:22.511073",
+  "course_id": "10363c5b-82b4-4121-8ef1-bae8fb42a5ce",
+  "content": [
+    {
+      "type": "x"
+    }
+  ],
+  "url_path": "/part-1/hello-world",
+  "title": "Hello world!",
+  "deleted_at": "2021-04-28T16:33:42.670935",
+  "chapter_id": "2495ffa3-7ea9-4615-baa5-828023688c79"
+}
+```
+*/
+#[instrument(skip(pool))]
+async fn delete_page(
+    request_page_id: web::Path<Uuid>,
+    pool: web::Data<PgPool>,
+    user: AuthUser,
+) -> ControllerResult<Json<Page>> {
+    let mut conn = pool.acquire().await?;
+    let deleted_page =
+        crate::models::pages::delete_page_and_exercises(&mut conn, *request_page_id).await?;
+    Ok(Json(deleted_page))
+}
+
+/**
+Add a route for each controller in this module.
+
+The name starts with an underline in order to appear before other functions in the module documentation.
+
+We add the routes by calling the route method instead of using the route annotations because this method preserves the function signatures for documentation.
+*/
+pub fn _add_pages_routes(cfg: &mut ServiceConfig) {
+    cfg.route("", web::post().to(post_new_page))
+        .route("/{page_id}", web::delete().to(delete_page));
+}
