+++ services/main-frontend/src/components/ChapterImageWidget.tsx
